refactor: Update storage server and backup configuration tables
This commit updates the storage server and backup configuration tables in the database migrations. It changes the data type of the `total_space_used` and `total_size` columns from `integer` to `bigInteger` to support larger values. This change ensures that the storage space usage and backup sizes can be accurately tracked and stored in the database.

+++ backend/database/migrations/2024_06_01_171934_create_backup_configurations_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->json('encryption_config');
             $table->json('integrity_check_config');
             $table->integer('total_backups')->default(0);
-            $table->integer('total_size')->default(0);
+            $table->bigInteger('total_size')->default(0);
             $table->dateTime('last_backup_at')->nullable();
             $table->integer('status')->default(0);
             $table->timestamps();
